add styling for ChatEditor
