--- pho/src/components/ChatContainer/ChatEditor/ChatEditor.css
@@ -1,42 +1,0 @@
-.content__footer {
-  padding-top: 30px;
-}
-.sendNewMessage {
-  background-color: #fff;
-  /* display: flex; */
-  justify-content: space-between;
-  padding: 10px;
-  border-radius: 8px;
-}
-.sendNewMessage button {
-  width: 100px;
-  height: 36px;
-  background-color: #ecefff;
-  border: none;
-  box-shadow: none;
-  outline: none;
-  cursor: pointer;
-  font-size: 16px;
-  color: #4665ff;
-  padding: 0;
-  border-radius: 5px;
-  line-height: 36px;
-  transition: all 0.3s cubic-bezier(0.88, 0.19, 0.37, 1.11);
-}
-.sendNewMessage button:hover {
-  transform: scale(1.2);
-}
-.sendNewMessage button i {
-  display: block;
-}
-.sendNewMessage input {
-  flex-grow: 1;
-  padding: 0 15px;
-  background-color: transparent;
-  border: none;
-  outline: none;
-}
-#sendMsgBtn {
-  background-color: #3b5bfe;
-  color: #fff;
-}
